+++ src/components/index.ts
@@ -1,4 +1,4 @@
 export { default as Hero } from './Hero';
 export { default as Location } from './Location';
-export { default as Navbar } from './Navbar';
+export { default as Navbar } from './navbar';
 export { default as Renderer } from './render';
